==> 02_Density.Rmd <==
---
title: "02_Density"
author: "Camille and Tristan"
date: "3/14/2023"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r Import data and libraries}
#Import relevant libraries
library(Distance); library(tidyverse); library(lme4); library(lmerTest); library(rsq); library(ggplot2);library(REdaS)

#Import relevant data 
transect_lengths <- read.csv("~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/Transect_Length.csv")

transect_lengths$Site_Transect <- paste(transect_lengths$Site,"_",as.numeric(transect_lengths$Transect), sep="")

survey_all <- read.csv("~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/lemurs_primary_in_transect_feb6.csv")
survey_all$Type[survey_all$Type=="diurnal"] <- "Diurnal"
survey_all$Site_Transect <- paste(survey_all$Site,"_",as.numeric(survey_all$Transect), sep="")

#NAs CURRENTLY BEING EXCLUDED, to include them set is.na(Dis) to 0
survey_all$Angle[is.na(survey_all$Angle)] <- pi/2
survey_all$distance <- abs((sin(survey_all$Angle))*(survey_all$Dis))

#remove detections not based on sight
survey_all <- survey_all %>% filter(ObjectID != 1440 & ObjectID != 1418 & ObjectID != 417 & ObjectID != 418 & ObjectID != 96)


#size as in group size - take the larger value of n_ind or sum of n_juv +n_ad
survey_all$N_ind <- as.numeric(survey_all$N_ind)
survey_all$N_ind[is.na(survey_all$N_ind)] <- 0
survey_all$N_ad[is.na(survey_all$N_ad)] <- 0
survey_all$N_juv[is.na(survey_all$N_juv)] <- 0
survey_all$N_total <- survey_all$N_ad + survey_all$N_juv
survey_all <- survey_all %>% mutate(N_total = ifelse(N_total > N_ind, N_total, N_ind))


#Nom_scientifique, Site_transect.x, Type, N_total
#sum1 <- survey_all %>% group_by(Site, Transect, Nom_sientifique,Type) %>% summarise(N_total = sum(as.numeric(N_total)))
#sum_noct <- sum1 %>% filter(Type=="Nocturnal") %>% pivot_wider(names_from = Nom_sientifique, values_from = N_total)
#sum_diur <- sum1 %>% filter(Type=="Diurnal") %>% pivot_wider(names_from = Nom_sientifique, values_from = N_total)

#NOBS - calculate before filtering by date
nobs <- survey_all %>% select(Site, Transect, Nom_sientifique, Type)%>%drop_na(Nom_sientifique) %>%group_by(Site, Transect, Nom_sientifique, Type) %>% mutate(NObs =n())%>% distinct()
nobs <- pivot_wider(nobs, names_from = Type, values_from = NObs)
nobs[is.na(nobs)]<- 0
nobs$NObs_total <- nobs$Diurnal + nobs$Nocturnal
nobs <- nobs %>% rename(Nobs_Diurnal = Diurnal)
nobs <- nobs %>% rename(Nobs_Nocturnal = Nocturnal)
nobs <-nobs %>% unite("Site_Transect", Site:Transect, sep="_")

nobs_cheiro <- survey_all%>%filter(Mois==1|Mois==2|Mois==3|Mois==11|Mois==12)%>% filter(Nom_sientifique=="Cheirogaleus_crossleyi") %>% select(Site, Transect, Nom_sientifique) %>%group_by(Site, Transect) %>% mutate(NObs =n())%>% distinct()
nobs_cheiro$Transect2 <- nobs_cheiro$Transect
nobs_cheiro$Site2 <- nobs_cheiro$Site
nobs_cheiro <-nobs_cheiro %>% unite("Site_Transect", Site2:Transect2, sep="_")


```

```{r Number of repetitions, number of observations, effort}
#NREP NOCTURNAL - Counts ALL REPS
nrep_nocturnal <- survey_all %>% select(Site, Transect, Repetition, Type)%>%filter(Type=="Nocturnal") %>%drop_na(Transect) %>% group_by(Site, Transect) %>% mutate(NRep = length(unique(Repetition))) %>% select(-Repetition)%>% distinct()
nrep_nocturnal$Site2 <- nrep_nocturnal$Site
nrep_nocturnal$Transect2 <- nrep_nocturnal$Transect
nrep_nocturnal <-nrep_nocturnal %>% unite("Site_Transect", Site2:Transect2, sep="_")

#NREP DIURNAL- Counts ALL REPS
nrep_diurnal <- survey_all %>% select(Site, Transect, Repetition, Type)%>%filter(Type=="Diurnal") %>%drop_na(Transect) %>% group_by(Site, Transect) %>% mutate(NRep = length(unique(Repetition))) %>% select(-Repetition)%>% distinct()
nrep_diurnal$Site2 <- nrep_diurnal$Site
nrep_diurnal$Transect2 <- nrep_diurnal$Transect
nrep_diurnal <-nrep_diurnal %>% unite("Site_Transect", Site2:Transect2, sep="_")

#CHEIRO- Counts ALL REPS
nrep_cheiro <- survey_all %>%filter(Mois==1|Mois==2|Mois==3|Mois==11|Mois==12) %>%filter(Type=="Nocturnal") %>%select(Site, Transect, Repetition)%>%drop_na(Transect) %>% group_by(Site, Transect) %>% mutate(NRep = length(unique(Repetition))) %>% select(-Repetition)%>% distinct()
nrep_cheiro$Site2 <- nrep_cheiro$Site
nrep_cheiro$Transect2 <- nrep_cheiro$Transect
nrep_cheiro <-nrep_cheiro %>% unite("Site_Transect", Site2:Transect2, sep="_")



#FILTER DATA FOR DISTANCE SAMPLING
survey_all$Date <- mdy(survey_all$Date)
survey_all_dist <- survey_all[survey_all$Date >= ymd("20220601"),]

#Make separate data frames for dirunal and nocturnal lemurs
survey_diurnal <- survey_all_dist %>% filter(Type=="Diurnal")
survey_nocturnal <- survey_all_dist %>% filter(Type=="Nocturnal")



#NREP NOCTURNAL - Only Reps with distance data
nrep_nocturnal_distance <- survey_all_dist %>% select(Site, Transect, Repetition, Type)%>%filter(Type=="Nocturnal") %>%drop_na(Transect) %>% group_by(Site, Transect) %>% mutate(NRep = length(unique(Repetition))) %>% select(-Repetition)%>% distinct()
nrep_nocturnal_distance$Site2 <- nrep_nocturnal_distance$Site
nrep_nocturnal_distance$Transect2 <- nrep_nocturnal_distance$Transect
nrep_nocturnal_distance <-nrep_nocturnal_distance %>% unite("Site_Transect", Site2:Transect2, sep="_")

#NREP DIURNAL- Only Reps with distance data
nrep_diurnal_distance <- survey_all_dist %>% select(Site, Transect, Repetition, Type)%>%filter(Type=="Diurnal") %>%drop_na(Transect) %>% group_by(Site, Transect) %>% mutate(NRep = length(unique(Repetition))) %>% select(-Repetition)%>% distinct()
nrep_diurnal_distance$Site2 <- nrep_diurnal_distance$Site
nrep_diurnal_distance$Transect2 <- nrep_diurnal_distance$Transect
nrep_diurnal_distance <-nrep_diurnal_distance %>% unite("Site_Transect", Site2:Transect2, sep="_")

#CHEIRO- Only Reps with distance data
nrep_cheiro_distance <- survey_all_dist %>%filter(Mois==1|Mois==2|Mois==3|Mois==11|Mois==12) %>%filter(Type=="Nocturnal" ) %>%select(Site, Transect, Repetition)%>%drop_na(Transect) %>% group_by(Site, Transect) %>% mutate(NRep = length(unique(Repetition))) %>% select(-Repetition)%>% distinct()
nrep_cheiro_distance$Site2 <- nrep_cheiro_distance$Site
nrep_cheiro_distance$Transect2 <- nrep_cheiro_distance$Transect
nrep_cheiro_distance <-nrep_cheiro_distance %>% unite("Site_Transect", Site2:Transect2, sep="_")


#EFFORT - only adding effort relevant to distance to the files, can calculate total with full version at end
survey_nocturnal <- left_join(survey_nocturnal, transect_lengths , by ="Site_Transect")
survey_nocturnal <- left_join(survey_nocturnal, nrep_nocturnal_distance, by ="Site_Transect")
survey_nocturnal$Effort.distance <- survey_nocturnal$Length* survey_nocturnal$NRep
survey_nocturnal <- left_join(survey_nocturnal, nrep_nocturnal, by ="Site_Transect")
survey_nocturnal$Effort.total <- survey_nocturnal$Length* survey_nocturnal$NRep.y
survey_nocturnal <- survey_nocturnal %>% select(-Type, - Type.y, -Site.y.y, -Transect.x.x, -Site.x.x, -Site.y, -Transect.y.y, -Transect.y)
survey_nocturnal <- survey_nocturnal %>% rename(Type = Type.x, Site = Site.x, Transect = Transect.x)
  
survey_diurnal <- left_join(survey_diurnal, transect_lengths, by ="Site_Transect")
survey_diurnal <- left_join(survey_diurnal, nrep_diurnal_distance, by ="Site_Transect")
survey_diurnal$Effort.distance <- survey_diurnal$Length* survey_diurnal$NRep
survey_diurnal <- left_join(survey_diurnal, nrep_diurnal, by ="Site_Transect")
survey_diurnal$Effort.total <- survey_diurnal$Length* survey_diurnal$NRep.y
survey_diurnal <- survey_diurnal %>% select(-Type, - Type.y, -Site.y.y, -Transect.x.x, -Site.x.x, -Site.y, -Transect.y.y, -Transect.y)
survey_diurnal <- survey_diurnal %>% rename(Type = Type.x, Site = Site.x, Transect = Transect.x)

survey_cheiro <- left_join((survey_all_dist %>% filter(Nom_sientifique=="Cheirogaleus_crossleyi" & (Mois >= 11 | Mois <=3 ))), transect_lengths , by ="Site_Transect")
survey_cheiro <- left_join(survey_cheiro, nrep_cheiro_distance, by ="Site_Transect")
survey_cheiro$Effort.distance <- survey_cheiro$Length* survey_cheiro$NRep
survey_cheiro <- left_join(survey_cheiro, nrep_cheiro, by ="Site_Transect")
survey_cheiro$Effort.total <- survey_cheiro$Length* survey_cheiro$NRep.y
survey_cheiro <- survey_cheiro %>% select( -Site.y.y, -Transect.x.x, -Site.x.x, -Site.y, -Transect.y.y, -Transect.y)
survey_cheiro <- survey_cheiro %>% rename( Site = Site.x, Transect = Transect.x)
```

```{r Allocebus density calculations}
#Get data frame for just allocebus
allocebus_df <- survey_nocturnal  %>% filter(Nom_sientifique=="Allocebus_trichotis")

#NAs currently being excluded
#allocebus_df$distance[is.na(allocebus_df$distance)] <- 0

#clean data data set
allocebus_df3 <- allocebus_df
allocebus_df3$Area <- 1000000 #For now, assume an area of 1... million 
allocebus_df3$Sample.Label <- 1
allocebus_df3$Region.Label <- allocebus_df3$Site_Transect
allocebus_df3$size <- as.numeric(allocebus_df3$N_total)
allocebus_df3$Effort <- allocebus_df3$Effort.distance


#colnames(allocebus_df3) <- c("Region.Label",  "Effort", "distance","Area", "Sample.Label")

hist(allocebus_df3$distance)

#ds_allocebus <- ds(allocebus_df3) - used for finding truncation dist
#check if group size affects detection dist
#ggplot(allocebus_df3, aes(distance,size))+geom_point()+geom_smooth(method=lm)
#ds_allocebus <- ds(allocebus_df3, truncation = 20, formula = ~size)

allocebus_truncation_dist <- 20
ds_allocebus <- ds(allocebus_df3, truncation = allocebus_truncation_dist)
gof_ds(ds_allocebus)


#Detection function models
model1_allocebus <- ds(allocebus_df3, key="hn", adjustment=NULL)
model2_allocebus <- ds(allocebus_df3, key="hn", adjustment="cos")
model3_allocebus <- ds(allocebus_df3, key="hn", adjustment="herm")
model4_allocebus <- ds(allocebus_df3, key="hr", adjustment=NULL)
model5_allocebus <- ds(allocebus_df3, key="hr", adjustment="cos")
model6_allocebus <- ds(allocebus_df3, key="hr", adjustment="herm")
model7_allocebus <- ds(allocebus_df3, key="unif", adjustment=NULL)
model8_allocebus <- ds(allocebus_df3, key="unif", adjustment="cos")
model9_allocebus <- ds(allocebus_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_allocebus)#significant, poor fit
gof_ds(model2_allocebus)
gof_ds(model3_allocebus)#significant, poor fit
gof_ds(model4_allocebus)
gof_ds(model5_allocebus)
gof_ds(model6_allocebus)
gof_ds(model7_allocebus)#significant, poor fit
gof_ds(model8_allocebus)
gof_ds(model9_allocebus)#significant, poor fit

#Of the models without poor fit, compare AIC calues
AIC(model2_allocebus, model4_allocebus,model5_allocebus, model6_allocebus,model8_allocebus)

#Plot the best-fitting model 
plot(model2_allocebus, nc=6)

#Get detection probability
allocebus_detection_prob <- predict(ds_allocebus)$fitted[[1]]

#Calculate strp width (2 * maximum distance, of the truncated distance data)
allocebus_strip_width <- allocebus_truncation_dist
#allocebus_strip_width <- max(allocebus_df3$distance) * 2

#Upload covariate data 

#NOTE THAT ALL DETECTIONS BEYOND TRUNCATION DISTANCE ARE BEING EXCLUDED
#CURRENTLY INCLUDING NA DISTANCE VALUES AS DETECTIONS
site_transects <- survey_all %>% group_by(Site_Transect) %>% summarise()

allocebus_df_old <- survey_all %>% filter(Nom_sientifique=="Allocebus_trichotis" & Type=="Nocturnal" & Date < ymd("20220601") & (Dis <1.56*allocebus_truncation_dist | is.na(Dis)))
allocebus_df_new <- survey_all %>% filter(Nom_sientifique=="Allocebus_trichotis" & Type=="Nocturnal"  & Date >= ymd("20220601") & (distance <= allocebus_truncation_dist  | is.na(distance)) )
allocebus_dat1 <- rbind(allocebus_df_new, allocebus_df_old)

allocebus_dat <- allocebus_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
allocebus_dat <- left_join(site_transects, allocebus_dat, by =c("Site_Transect"))

nobs_allocebus <- allocebus_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
allocebus_dat <- left_join(allocebus_dat, nobs_allocebus, by =c("Site_Transect"))
allocebus_dat[is.na(allocebus_dat)] <- 0

allocebus_dat <- left_join(allocebus_dat,nrep_nocturnal, by =c("Site_Transect"))
allocebus_dat <- allocebus_dat %>% select(-Site, -Transect, -Type)

allocebus_dat <- left_join(allocebus_dat,transect_lengths, by =c("Site_Transect"))
allocebus_dat$Density <- (allocebus_dat$Nind/allocebus_detection_prob)/(allocebus_truncation_dist*allocebus_dat$Length*allocebus_dat$NRep)
allocebus_dat$Density_km2 <- allocebus_dat$Density*1000000

write.csv(allocebus_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/allocebus_dat.csv")

#allocebus_dat <- read.csv("~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/allo_reps.csv")

#allocebus_dat <- left_join(allocebus_dat, nrep_nocturnal, by =c("Site_Transect"))
#allocebus_dat <- left_join(allocebus_dat, nobs[nobs$Nom_sientifique=="Allocebus_trichotis",], by =c("Site_Transect"))
#allocebus_dat <- allocebus_dat %>% select(Site_Transect, NObs, NRep)
#allocebus_dat$NObs[is.na(allocebus_dat$NObs)] <- 0

#Use global detection function to calulate density at each transect
#allocebus_dat$M_surveyed <- allocebus_dat$NRep *1000*  allocebus_strip_width
#allocebus_dat$Density <- (allocebus_dat$NObs) / (allocebus_dat$M_surveyed*allocebus_detection_prob) * 1000000
#allocebus_dat <- allocebus_dat %>% select(Site_Transect, Density)

#Clean transect names
covs_allo$Site_Transect[covs_allo$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_allo$Site_Transect[covs_allo$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_allo$Site_Transect[covs_allo$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

#Join density data to covariate data 
covs_allo_density <- left_join(allocebus_dat, covs_allo, by="Site_Transect")
covs_allo_density$Species <- "Allocebus_trichotis"

#Linear model of density as explained by combinations of the covariates
lmer_allocebus <- lmer(data=covs_allo_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_allocebus1 <- lmer(data=covs_allo_density , Density~ elevation +(1|Site))
lmer_allocebus2 <- lmer(data=covs_allo_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_allocebus3 <- lmer(data=covs_allo_density , Density~  height + (1|Site))
lmer_allocebus4 <- lmer(data=covs_allo_density , Density~ rug +(1|Site))
lmer_allocebus5 <- lmer(data=covs_allo_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_allocebus6 <- lmer(data=covs_allo_density , Density~ elevation + height +(1|Site))
lmer_allocebus7 <- lmer(data=covs_allo_density , Density~ elevation +  rug +(1|Site))
lmer_allocebus8 <- lmer(data=covs_allo_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_allocebus9 <- lmer(data=covs_allo_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_allocebus10 <- lmer(data=covs_allo_density , Density~ height + rug +(1|Site))
lmer_allocebus11 <- lmer(data=covs_allo_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_allocebus12 <- lmer(data=covs_allo_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_allocebus13 <- lmer(data=covs_allo_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_allocebus14 <- lmer(data=covs_allo_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_allocebus_rsq <- rsq.lmm(lmer_allocebus)[[1]]
lmer_allocebus_rsq1 <- rsq.lmm(lmer_allocebus1)[[1]]
lmer_allocebus_rsq2 <- rsq.lmm(lmer_allocebus2)[[1]]
lmer_allocebus_rsq3 <- rsq.lmm(lmer_allocebus3)[[1]]
lmer_allocebus_rsq4 <- rsq.lmm(lmer_allocebus4)[[1]]
lmer_allocebus_rsq5 <- rsq.lmm(lmer_allocebus5)[[1]]
lmer_allocebus_rsq6 <- rsq.lmm(lmer_allocebus6)[[1]]
lmer_allocebus_rsq7 <- rsq.lmm(lmer_allocebus7)[[1]]
lmer_allocebus_rsq8 <- rsq.lmm(lmer_allocebus8)[[1]]
lmer_allocebus_rsq9 <- rsq.lmm(lmer_allocebus9)[[1]]
lmer_allocebus_rsq10 <- rsq.lmm(lmer_allocebus10)[[1]]
lmer_allocebus_rsq11<- rsq.lmm(lmer_allocebus11)[[1]]
lmer_allocebus_rsq12 <- rsq.lmm(lmer_allocebus12)[[1]]
lmer_allocebus_rsq13 <- rsq.lmm(lmer_allocebus13)[[1]]
lmer_allocebus_rsq14 <- rsq.lmm(lmer_allocebus14)[[1]]
max(lmer_allocebus_rsq, lmer_allocebus_rsq1 , lmer_allocebus_rsq2 , lmer_allocebus_rsq3 , lmer_allocebus_rsq4 , lmer_allocebus_rsq5 , lmer_allocebus_rsq6, lmer_allocebus_rsq7 , lmer_allocebus_rsq8 , lmer_allocebus_rsq9 , lmer_allocebus_rsq10, lmer_allocebus_rsq11 , lmer_allocebus_rsq12 , lmer_allocebus_rsq13 , lmer_allocebus_rsq14 )

lmer_allocebus_best <-lmer_allocebus9
summary(lmer_allocebus_best)

```

```{r Microcebus density calculations}

#Get data frame for just microcebus
microcebus_df <- survey_nocturnal  %>% filter(Nom_sientifique=="Microcebus_lehilahytsara")

#microcebus_df$distance[is.na(microcebus_df$distance)] <- 0

#Clean data frame
microcebus_df3 <- microcebus_df
microcebus_df3$Area <- 1000000 #For now, assume an area of 1... million 
microcebus_df3$Sample.Label <- 1
microcebus_df3$Region.Label <- microcebus_df3$Site_Transect
microcebus_df3$size <- as.numeric(microcebus_df3$N_total)
microcebus_df3$Effort <- microcebus_df3$Effort.distance

#Detection function models
microcebus_truncation_dist <- 15
ds_microcebus <- ds(microcebus_df3, truncation = microcebus_truncation_dist)
gof_ds(ds_microcebus) #poor fit, significant

model1_microcebus <- ds(microcebus_df3, key="hn", adjustment=NULL)
model2_microcebus <- ds(microcebus_df3, key="hn", adjustment="cos")
model3_microcebus <- ds(microcebus_df3, key="hn", adjustment="herm")
model4_microcebus <- ds(microcebus_df3, key="hr", adjustment=NULL)
model5_microcebus <- ds(microcebus_df3, key="hr", adjustment="cos")
model6_microcebus <- ds(microcebus_df3, key="hr", adjustment="herm")
model7_microcebus <- ds(microcebus_df3, key="unif", adjustment=NULL)
model8_microcebus <- ds(microcebus_df3, key="unif", adjustment="cos")
model9_microcebus <- ds(microcebus_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_microcebus)#significant 
gof_ds(model2_microcebus)#significant 
gof_ds(model3_microcebus)#significant 
gof_ds(model4_microcebus)#significant 
gof_ds(model5_microcebus)#significant 
gof_ds(model6_microcebus)#significant
gof_ds(model7_microcebus)#significant
gof_ds(model8_microcebus)#significant
gof_ds(model9_microcebus)#significant

#Of the models without poor fit, compare AIC calues
AIC(model5_microcebus)

#Plot the best-fitting model 
plot(model5_microcebus, nc=6)

#Get detection probability
microcebus_detection_prob <- predict(ds_microcebus)$fitted[[1]]

#Calculate strp width (2 * maximum distance, of the truncated distance data)
microcebus_strip_width <- microcebus_truncation_dist

#Upload covariate data 
microcebus_df_old <- survey_all %>% filter(Nom_sientifique=="Microcebus_lehilahytsara" & Type=="Nocturnal" & Date < ymd("20220601") & (Dis <1.56*microcebus_truncation_dist | is.na(Dis)))
microcebus_df_new <- survey_all %>% filter(Nom_sientifique=="Microcebus_lehilahytsara" & Type=="Nocturnal"  & Date >= ymd("20220601") & (distance <= microcebus_truncation_dist  | is.na(distance)) )
microcebus_dat1 <- rbind(microcebus_df_new, microcebus_df_old)

microcebus_dat <- microcebus_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
microcebus_dat <- left_join(site_transects, microcebus_dat, by =c("Site_Transect"))

nobs_microcebus <- microcebus_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
microcebus_dat <- left_join(microcebus_dat, nobs_microcebus, by =c("Site_Transect"))
microcebus_dat[is.na(microcebus_dat)] <- 0

microcebus_dat <- left_join(microcebus_dat,nrep_nocturnal, by =c("Site_Transect"))
microcebus_dat <- microcebus_dat %>% select(-Site, -Transect, -Type)

microcebus_dat <- left_join(microcebus_dat,transect_lengths, by =c("Site_Transect"))
microcebus_dat$Density <- (microcebus_dat$Nind/microcebus_detection_prob)/(microcebus_truncation_dist*microcebus_dat$Length*microcebus_dat$NRep)
microcebus_dat$Density_km2 <- microcebus_dat$Density*1000000

write.csv(microcebus_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/microcebus_dat.csv")


#microcebus_dat <- read.csv("~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/micro_reps.csv")

#microcebus_dat <- left_join(microcebus_dat, nrep_nocturnal, by =c("Site_Transect"))
#microcebus_dat <- left_join(microcebus_dat, #nobs[nobs$Nom_sientifique=="Microcebus_lehilahytsara",], by =c("Site_Transect"))
#microcebus_dat <- microcebus_dat %>% select(Site_Transect, NObs, NRep)
#microcebus_dat$NObs[is.na(microcebus_dat$NObs)] <- 0

#Use global detection function to calulate density at each transect
#microcebus_dat$M_surveyed <- microcebus_dat$NRep *1000*  microcebus_strip_width
#microcebus_dat$Density <- (microcebus_dat$NObs) / (microcebus_dat$M_surveyed*microcebus_detection_prob) * 1000000
#microcebus_dat <- microcebus_dat %>% select(Site_Transect, Density)

#Clean transect names
covs_micro$Site_Transect[covs_micro$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_micro$Site_Transect[covs_micro$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_micro$Site_Transect[covs_micro$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

#Join density data to covariate data 
covs_micro_density <- left_join(microcebus_dat, covs_micro, by="Site_Transect")
covs_micro_density$Species <- "Microcebus_mittermeiri"

#Linear model of density as explained by combinations of the covariates
lmer_microcebus <- lmer(data=covs_micro_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_microcebus1 <- lmer(data=covs_micro_density , Density~ elevation +(1|Site))
lmer_microcebus2 <- lmer(data=covs_micro_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_microcebus3 <- lmer(data=covs_micro_density , Density~  height + (1|Site))
lmer_microcebus4 <- lmer(data=covs_micro_density , Density~ rug +(1|Site))
lmer_microcebus5 <- lmer(data=covs_micro_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_microcebus6 <- lmer(data=covs_micro_density , Density~ elevation + height +(1|Site))
lmer_microcebus7 <- lmer(data=covs_micro_density , Density~ elevation +  rug +(1|Site))
lmer_microcebus8 <- lmer(data=covs_micro_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_microcebus9 <- lmer(data=covs_micro_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_microcebus10 <- lmer(data=covs_micro_density , Density~ height + rug +(1|Site))
lmer_microcebus11 <- lmer(data=covs_micro_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_microcebus12 <- lmer(data=covs_micro_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_microcebus13 <- lmer(data=covs_micro_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_microcebus14 <- lmer(data=covs_micro_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_microcebus_rsq <- rsq.lmm(lmer_microcebus)[[1]]
lmer_microcebus_rsq1 <- rsq.lmm(lmer_microcebus1)[[1]]
lmer_microcebus_rsq2 <- rsq.lmm(lmer_microcebus2)[[1]]
lmer_microcebus_rsq3 <- rsq.lmm(lmer_microcebus3)[[1]]
lmer_microcebus_rsq4 <- rsq.lmm(lmer_microcebus4)[[1]]
lmer_microcebus_rsq5 <- rsq.lmm(lmer_microcebus5)[[1]]
lmer_microcebus_rsq6 <- rsq.lmm(lmer_microcebus6)[[1]]
lmer_microcebus_rsq7 <- rsq.lmm(lmer_microcebus7)[[1]]
lmer_microcebus_rsq8 <- rsq.lmm(lmer_microcebus8)[[1]]
lmer_microcebus_rsq9 <- rsq.lmm(lmer_microcebus9)[[1]]
lmer_microcebus_rsq10 <- rsq.lmm(lmer_microcebus10)[[1]]
lmer_microcebus_rsq11<- rsq.lmm(lmer_microcebus11)[[1]]
lmer_microcebus_rsq12 <- rsq.lmm(lmer_microcebus12)[[1]]
lmer_microcebus_rsq13 <- rsq.lmm(lmer_microcebus13)[[1]]
lmer_microcebus_rsq14 <- rsq.lmm(lmer_microcebus14)[[1]]
max(lmer_microcebus_rsq, lmer_microcebus_rsq1 , lmer_microcebus_rsq2 , lmer_microcebus_rsq3 , lmer_microcebus_rsq4 , lmer_microcebus_rsq5 , lmer_microcebus_rsq6, lmer_microcebus_rsq7 , lmer_microcebus_rsq8 , lmer_microcebus_rsq9 , lmer_microcebus_rsq10, lmer_microcebus_rsq11 , lmer_microcebus_rsq12 , lmer_microcebus_rsq13 , lmer_microcebus_rsq14 )

lmer_microcebus_best <- lmer_microcebus8
summary(lmer_microcebus_best)

```

```{r Cheirogaleus density calculations: NOT ALL SITES}
#Get data frame for just cheirogaleus
cheirogaleus_df <- survey_cheiro

#cheirogaleus_df <- left_join(cheirogaleus_df, nrep_cheiro, by =c("Site", "Transect"))
#cheirogaleus_df$Effort <- cheirogaleus_df$Length *cheirogaleus_df$NRep

#Calculate perpendicular distance from angle and distance
#cheirogaleus_df$distance[is.na(cheirogaleus_df$distance)] <- 0


#Clean data frame
cheirogaleus_df3 <- cheirogaleus_df
cheirogaleus_df3$Area <- 1000000 #For now, assume an area of 1... million 
cheirogaleus_df3$Sample.Label <- 1
cheirogaleus_df3$Region.Label <- cheirogaleus_df3$Site_Transect
cheirogaleus_df3$size <- as.numeric(cheirogaleus_df3$N_total)
cheirogaleus_df3$Effort <- cheirogaleus_df3$Effort.distance


#Detection function models
cheirogaleus_truncation_dist <- 25
ds_cheirogaleus <- ds(cheirogaleus_df3, truncation = cheirogaleus_truncation_dist)
gof_ds(ds_cheirogaleus) #significant, poor fit

model1_cheirogaleus <- ds(cheirogaleus_df3, key="hn", adjustment=NULL)
model2_cheirogaleus <- ds(cheirogaleus_df3, key="hn", adjustment="cos")
model3_cheirogaleus <- ds(cheirogaleus_df3, key="hn", adjustment="herm")
model4_cheirogaleus <- ds(cheirogaleus_df3, key="hr", adjustment=NULL)
model5_cheirogaleus <- ds(cheirogaleus_df3, key="hr", adjustment="cos")
model6_cheirogaleus <- ds(cheirogaleus_df3, key="hr", adjustment="herm")
model7_cheirogaleus <- ds(cheirogaleus_df3, key="unif", adjustment=NULL)
model8_cheirogaleus <- ds(cheirogaleus_df3, key="unif", adjustment="cos")
model9_cheirogaleus <- ds(cheirogaleus_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_cheirogaleus)#significant
gof_ds(model2_cheirogaleus)#significant
gof_ds(model3_cheirogaleus)#significant
gof_ds(model4_cheirogaleus)#significant
gof_ds(model5_cheirogaleus)#significant
gof_ds(model6_cheirogaleus)#significant
gof_ds(model7_cheirogaleus)#significant
gof_ds(model8_cheirogaleus)#significant
gof_ds(model9_cheirogaleus)#significant

#all have poor fit :( 

#compare AIC calues
AIC(model1_cheirogaleus, model2_cheirogaleus, model3_cheirogaleus, model4_cheirogaleus,model5_cheirogaleus, model6_cheirogaleus,model7_cheirogaleus,  model8_cheirogaleus, model9_cheirogaleus)

#Plot the best-fitting model 
plot(model4_cheirogaleus, nc=4)

#Get detection probability
cheirogaleus_detection_prob <- predict(ds_cheirogaleus)$fitted[[1]]

#Calculate strp width (2 * maximum distance, of the truncated distance data)
cheirogaleus_strip_width <- cheirogaleus_truncation_dist

#Upload covariate data 
cheirogaleus_df_old <- survey_all %>% filter(Nom_sientifique=="Cheirogaleus_crossleyi" & Type=="Nocturnal" & Date < ymd("20220601") & (Mois >= 11 | Mois <=3 ) & (Dis <1.56*cheirogaleus_truncation_dist | is.na(Dis)))
cheirogaleus_df_new <- survey_all %>% filter(Nom_sientifique=="Cheirogaleus_crossleyi" & Type=="Nocturnal"  & Date >= ymd("20220601" )& (Mois >= 11 | Mois <=3 ) & (distance <= cheirogaleus_truncation_dist  | is.na(distance)) )
cheirogaleus_dat1 <- rbind(cheirogaleus_df_new, cheirogaleus_df_old)

cheirogaleus_dat <- cheirogaleus_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
cheirogaleus_dat <- left_join(site_transects, cheirogaleus_dat, by =c("Site_Transect"))

nobs_cheirogaleus <- cheirogaleus_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
cheirogaleus_dat <- left_join(cheirogaleus_dat, nobs_cheirogaleus, by =c("Site_Transect"))
cheirogaleus_dat[is.na(cheirogaleus_dat)] <- 0

cheirogaleus_dat <- left_join(cheirogaleus_dat,nrep_cheiro, by =c("Site_Transect"))
cheirogaleus_dat <- cheirogaleus_dat %>% select(-Site, -Transect)

cheirogaleus_dat <- left_join(cheirogaleus_dat,transect_lengths, by =c("Site_Transect"))
cheirogaleus_dat$Density <- (cheirogaleus_dat$Nind/cheirogaleus_detection_prob)/(cheirogaleus_truncation_dist*cheirogaleus_dat$Length*cheirogaleus_dat$NRep)
cheirogaleus_dat$Density_km2 <- cheirogaleus_dat$Density*1000000

write.csv(cheirogaleus_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/cheirogaleus_dat.csv")

#Clean transect names
covs_cheiro$Site_Transect[covs_cheiro$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_cheiro$Site_Transect[covs_cheiro$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_cheiro$Site_Transect[covs_cheiro$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

#Join density data to covariate data 
covs_cheiro_density <- left_join(cheirogaleus_dat, covs_cheiro, by="Site_Transect")
covs_cheiro_density$Species <- "Cheirogaleus_crossleyi"

#Linear model of density as explained by combinations of the covariates
lmer_cheirogaleus <- lmer(data=covs_cheiro_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_cheirogaleus1 <- lmer(data=covs_cheiro_density , Density~ elevation +(1|Site))
lmer_cheirogaleus2 <- lmer(data=covs_cheiro_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_cheirogaleus3 <- lmer(data=covs_cheiro_density , Density~  height + (1|Site))
lmer_cheirogaleus4 <- lmer(data=covs_cheiro_density , Density~ rug +(1|Site))
lmer_cheirogaleus5 <- lmer(data=covs_cheiro_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_cheirogaleus6 <- lmer(data=covs_cheiro_density , Density~ elevation + height +(1|Site))
lmer_cheirogaleus7 <- lmer(data=covs_cheiro_density , Density~ elevation +  rug +(1|Site))
lmer_cheirogaleus8 <- lmer(data=covs_cheiro_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_cheirogaleus9 <- lmer(data=covs_cheiro_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_cheirogaleus10 <- lmer(data=covs_cheiro_density , Density~ height + rug +(1|Site))
lmer_cheirogaleus11 <- lmer(data=covs_cheiro_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_cheirogaleus12 <- lmer(data=covs_cheiro_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_cheirogaleus13 <- lmer(data=covs_cheiro_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_cheirogaleus14 <- lmer(data=covs_cheiro_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_cheirogaleus_rsq <- rsq.lmm(lmer_cheirogaleus)[[1]]
lmer_cheirogaleus_rsq1 <- rsq.lmm(lmer_cheirogaleus1)[[1]]
lmer_cheirogaleus_rsq2 <- rsq.lmm(lmer_cheirogaleus2)[[1]]
lmer_cheirogaleus_rsq3 <- rsq.lmm(lmer_cheirogaleus3)[[1]]
lmer_cheirogaleus_rsq4 <- rsq.lmm(lmer_cheirogaleus4)[[1]]
lmer_cheirogaleus_rsq5 <- rsq.lmm(lmer_cheirogaleus5)[[1]]
lmer_cheirogaleus_rsq6 <- rsq.lmm(lmer_cheirogaleus6)[[1]]
lmer_cheirogaleus_rsq7 <- rsq.lmm(lmer_cheirogaleus7)[[1]]
lmer_cheirogaleus_rsq8 <- rsq.lmm(lmer_cheirogaleus8)[[1]]
lmer_cheirogaleus_rsq9 <- rsq.lmm(lmer_cheirogaleus9)[[1]]
lmer_cheirogaleus_rsq10 <- rsq.lmm(lmer_cheirogaleus10)[[1]]
lmer_cheirogaleus_rsq11<- rsq.lmm(lmer_cheirogaleus11)[[1]]
lmer_cheirogaleus_rsq12 <- rsq.lmm(lmer_cheirogaleus12)[[1]]
lmer_cheirogaleus_rsq13 <- rsq.lmm(lmer_cheirogaleus13)[[1]]
lmer_cheirogaleus_rsq14 <- rsq.lmm(lmer_cheirogaleus14)[[1]]
max(lmer_cheirogaleus_rsq, lmer_cheirogaleus_rsq1 , lmer_cheirogaleus_rsq2 , lmer_cheirogaleus_rsq3 , lmer_cheirogaleus_rsq4 , lmer_cheirogaleus_rsq5 , lmer_cheirogaleus_rsq6, lmer_cheirogaleus_rsq7 , lmer_cheirogaleus_rsq8 , lmer_cheirogaleus_rsq9 , lmer_cheirogaleus_rsq10, lmer_cheirogaleus_rsq11 , lmer_cheirogaleus_rsq12 , lmer_cheirogaleus_rsq13 , lmer_cheirogaleus_rsq14 )

lmer_cheirogaleus_best <- lmer_cheirogaleus11 #forest loss proportion has a signifcantly positive relationship with Cheirogaleus density

```

```{r Lepilemur density calculations}

#Get data frame for just lepilemur
lepilemur_df <- survey_nocturnal  %>% filter(Nom_sientifique=="Lepilemur_seali")

#lepilemur_df$distance[is.na(lepilemur_df$distance)] <- 0

#Clean data frame
lepilemur_df3 <- lepilemur_df
lepilemur_df3$Area <- 1000000 #For now, assume an area of 1... million 
lepilemur_df3$Sample.Label <- 1
lepilemur_df3$Region.Label <- lepilemur_df3$Site_Transect
lepilemur_df3$size <- as.numeric(lepilemur_df3$N_total)
lepilemur_df3$Effort <- lepilemur_df3$Effort.distance

#Detection function models
lepilemur_truncation_dist <- 25
ds_lepilemur <- ds(lepilemur_df3, truncation = lepilemur_truncation_dist)

model1_lepilemur <- ds(lepilemur_df3, key="hn", adjustment=NULL)
model2_lepilemur <- ds(lepilemur_df3, key="hn", adjustment="cos")
model3_lepilemur <- ds(lepilemur_df3, key="hn", adjustment="herm")
model4_lepilemur <- ds(lepilemur_df3, key="hr", adjustment=NULL)
model5_lepilemur <- ds(lepilemur_df3, key="hr", adjustment="cos")
model6_lepilemur <- ds(lepilemur_df3, key="hr", adjustment="herm")
model7_lepilemur <- ds(lepilemur_df3, key="unif", adjustment=NULL)
model8_lepilemur <- ds(lepilemur_df3, key="unif", adjustment="cos")
model9_lepilemur <- ds(lepilemur_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_lepilemur)#significant
gof_ds(model2_lepilemur)
gof_ds(model3_lepilemur)#significant
gof_ds(model4_lepilemur)
gof_ds(model5_lepilemur)
gof_ds(model6_lepilemur)
gof_ds(model7_lepilemur)#significant
gof_ds(model8_lepilemur)
gof_ds(model9_lepilemur)#significant,

#Of the models without poor fit, compare AIC calues
AIC( model2_lepilemur, model4_lepilemur, model5_lepilemur,model6_lepilemur, model8_lepilemur)

#Plot the best-fitting model 
plot(model2_lepilemur, nc=4)

#Get detection probability
lepilemur_detection_prob <- predict(ds_lepilemur)$fitted[[1]]
gof_ds(ds_lepilemur)

#Calculate strp width (2 * maximum distance, of the truncated distance data)
lepilemur_strip_width <- lepilemur_truncation_dist

lepilemur_df_old <- survey_all %>% filter(Nom_sientifique=="Lepilemur_seali" & Type=="Nocturnal" & Date < ymd("20220601") & (Dis <1.56*lepilemur_truncation_dist | is.na(Dis)))
lepilemur_df_new <- survey_all %>% filter(Nom_sientifique=="Lepilemur_seali" & Type=="Nocturnal"  & Date >= ymd("20220601") & (distance <= lepilemur_truncation_dist  | is.na(distance)) )
lepilemur_dat1 <- rbind(lepilemur_df_new, lepilemur_df_old)

lepilemur_dat <- lepilemur_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
lepilemur_dat <- left_join(site_transects, lepilemur_dat, by =c("Site_Transect"))

nobs_lepilemur <- lepilemur_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
lepilemur_dat <- left_join(lepilemur_dat, nobs_lepilemur, by =c("Site_Transect"))
lepilemur_dat[is.na(lepilemur_dat)] <- 0

lepilemur_dat <- left_join(lepilemur_dat,nrep_nocturnal, by =c("Site_Transect"))
lepilemur_dat <- lepilemur_dat %>% select(-Site, -Transect, -Type)

lepilemur_dat <- left_join(lepilemur_dat,transect_lengths, by =c("Site_Transect"))
lepilemur_dat$Density <- (lepilemur_dat$Nind/lepilemur_detection_prob)/(lepilemur_truncation_dist*lepilemur_dat$Length*lepilemur_dat$NRep)
lepilemur_dat$Density_km2 <- lepilemur_dat$Density*1000000

write.csv(lepilemur_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/lepilemur_dat.csv")

#Upload covariate data 
#lepilemur_dat <- read.csv("~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/lepi_reps.csv")

#lepilemur_dat <- left_join(lepilemur_dat, nrep_nocturnal, by =c("Site_Transect"))
#lepilemur_dat <- left_join(lepilemur_dat, nobs[nobs$Nom_sientifique=="Lepilemur_seali",], by =c("Site_Transect"))
#lepilemur_dat <- lepilemur_dat %>% select(Site_Transect, NObs, NRep)
#lepilemur_dat$NObs[is.na(lepilemur_dat$NObs)] <- 0

#Use global detection function to calulate density at each transect
#lepilemur_dat$M_surveyed <- lepilemur_dat$NRep *1000*  lepilemur_strip_width
#lepilemur_dat$Density <- (lepilemur_dat$NObs) / (lepilemur_dat$M_surveyed*lepilemur_detection_prob) * 1000000
#lepilemur_dat <- lepilemur_dat %>% select(Site_Transect, Density)

#Clean transect names
covs_lepi$Site_Transect[covs_lepi$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_lepi$Site_Transect[covs_lepi$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_lepi$Site_Transect[covs_lepi$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

#Join density data to covariate data 
covs_lepi_density <- left_join(lepilemur_dat, covs_lepi, by="Site_Transect")
covs_lepi_density$Species <- "Lepilemur_seali"

#Linear model of density as explained by combinations of the covariates
lmer_lepilemur <- lmer(data=covs_lepi_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_lepilemur1 <- lmer(data=covs_lepi_density , Density~ elevation +(1|Site))
lmer_lepilemur2 <- lmer(data=covs_lepi_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_lepilemur3 <- lmer(data=covs_lepi_density , Density~  height + (1|Site))
lmer_lepilemur4 <- lmer(data=covs_lepi_density , Density~ rug +(1|Site))
lmer_lepilemur5 <- lmer(data=covs_lepi_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_lepilemur6 <- lmer(data=covs_lepi_density , Density~ elevation + height +(1|Site))
lmer_lepilemur7 <- lmer(data=covs_lepi_density , Density~ elevation +  rug +(1|Site))
lmer_lepilemur8 <- lmer(data=covs_lepi_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_lepilemur9 <- lmer(data=covs_lepi_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_lepilemur10 <- lmer(data=covs_lepi_density , Density~ height + rug +(1|Site))
lmer_lepilemur11 <- lmer(data=covs_lepi_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_lepilemur12 <- lmer(data=covs_lepi_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_lepilemur13 <- lmer(data=covs_lepi_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_lepilemur14 <- lmer(data=covs_lepi_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_lepilemur_rsq <- rsq.lmm(lmer_lepilemur)[[1]]
lmer_lepilemur_rsq1 <- rsq.lmm(lmer_lepilemur1)[[1]]
lmer_lepilemur_rsq2 <- rsq.lmm(lmer_lepilemur2)[[1]]
lmer_lepilemur_rsq3 <- rsq.lmm(lmer_lepilemur3)[[1]]
lmer_lepilemur_rsq4 <- rsq.lmm(lmer_lepilemur4)[[1]]
lmer_lepilemur_rsq5 <- rsq.lmm(lmer_lepilemur5)[[1]]
lmer_lepilemur_rsq6 <- rsq.lmm(lmer_lepilemur6)[[1]]
lmer_lepilemur_rsq7 <- rsq.lmm(lmer_lepilemur7)[[1]]
lmer_lepilemur_rsq8 <- rsq.lmm(lmer_lepilemur8)[[1]]
lmer_lepilemur_rsq9 <- rsq.lmm(lmer_lepilemur9)[[1]]
lmer_lepilemur_rsq10 <- rsq.lmm(lmer_lepilemur10)[[1]]
lmer_lepilemur_rsq11<- rsq.lmm(lmer_lepilemur11)[[1]]
lmer_lepilemur_rsq12 <- rsq.lmm(lmer_lepilemur12)[[1]]
lmer_lepilemur_rsq13 <- rsq.lmm(lmer_lepilemur13)[[1]]
lmer_lepilemur_rsq14 <- rsq.lmm(lmer_lepilemur14)[[1]]
max(lmer_lepilemur_rsq, lmer_lepilemur_rsq1 , lmer_lepilemur_rsq2 , lmer_lepilemur_rsq3 , lmer_lepilemur_rsq4 , lmer_lepilemur_rsq5 , lmer_lepilemur_rsq6, lmer_lepilemur_rsq7 , lmer_lepilemur_rsq8 , lmer_lepilemur_rsq9 , lmer_lepilemur_rsq10, lmer_lepilemur_rsq11 , lmer_lepilemur_rsq12 , lmer_lepilemur_rsq13 , lmer_lepilemur_rsq14 )

lmer_lepilemur_best <- lmer_lepilemur4 #rugedness has a significantly negative relationship with lepilemur density
summary(lmer_lepilemur4)

```

```{r Avahi density calculations}

#Get data frame for just avahi
avahi_df <- survey_nocturnal  %>% filter(Nom_sientifique=="Avahi_laniger")

#avahi_df$distance[is.na(avahi_df$distance)] <- 0

#Clean data frame
avahi_df3 <- avahi_df
avahi_df3$Area <- 1000000 #For now, assume an area of 1... million 
avahi_df3$Sample.Label <- 1
avahi_df3$Region.Label <- avahi_df3$Site_Transect
avahi_df3$size <- as.numeric(avahi_df3$N_total)
avahi_df3$Effort <- avahi_df3$Effort.distance


#Detection function models
avahi_truncation_dist <- 25
ds_avahi <- ds(avahi_df3, truncation = avahi_truncation_dist)
gof_ds(ds_avahi) #significant, poor fit
#

model1_avahi <- ds(avahi_df3, key="hn", adjustment=NULL)
model2_avahi <- ds(avahi_df3, key="hn", adjustment="cos")
model3_avahi <- ds(avahi_df3, key="hn", adjustment="herm")
model4_avahi <- ds(avahi_df3, key="hr", adjustment=NULL)
model5_avahi <- ds(avahi_df3, key="hr", adjustment="cos")
model6_avahi <- ds(avahi_df3, key="hr", adjustment="herm")
model7_avahi <- ds(avahi_df3, key="unif", adjustment=NULL)
model8_avahi <- ds(avahi_df3, key="unif", adjustment="cos")
model9_avahi <- ds(avahi_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_avahi)#significant
gof_ds(model2_avahi)#significant
gof_ds(model3_avahi)# significant
gof_ds(model4_avahi)#significant
gof_ds(model5_avahi)#significant
gof_ds(model6_avahi)#significant
gof_ds(model7_avahi)#significant
gof_ds(model8_avahi)#significant
gof_ds(model9_avahi)#significant

#all significant:(

#compare AIC calues
AIC(model1_avahi,model2_avahi,model3_avahi,model4_avahi,model5_avahi,model6_avahi,model7_avahi,model8_avahi, model9_avahi)

#Plot the best-fitting model 
plot(model4_avahi, nc=4)
#model 4 looks ridiculous so go with 2 instead
plot(model2_avahi, nc=4)

#Get detection probability
avahi_detection_prob <- predict(ds_avahi)$fitted[[1]]

#Calculate strp width (2 * maximum distance, of the truncated distance data)
avahi_strip_width <- allocebus_truncation_dist

#Upload covariate data 
avahi_df_old <- survey_all %>% filter(Nom_sientifique=="Avahi_laniger" & Type=="Nocturnal" & Date < ymd("20220601") & (Dis <1.56*avahi_truncation_dist | is.na(Dis)))
avahi_df_new <- survey_all %>% filter(Nom_sientifique=="Avahi_laniger" & Type=="Nocturnal"  & Date >= ymd("20220601") & (distance <= avahi_truncation_dist  | is.na(distance)) )
avahi_dat1 <- rbind(avahi_df_new, avahi_df_old)

avahi_dat <- avahi_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
avahi_dat <- left_join(site_transects, avahi_dat, by =c("Site_Transect"))

nobs_avahi <- avahi_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
avahi_dat <- left_join(avahi_dat, nobs_avahi, by =c("Site_Transect"))
avahi_dat[is.na(avahi_dat)] <- 0

avahi_dat <- left_join(avahi_dat,nrep_nocturnal, by =c("Site_Transect"))
avahi_dat <- avahi_dat %>% select(-Site, -Transect, -Type)

avahi_dat <- left_join(avahi_dat,transect_lengths, by =c("Site_Transect"))
avahi_dat$Density <- (avahi_dat$Nind/avahi_detection_prob)/(avahi_truncation_dist*avahi_dat$Length*avahi_dat$NRep)
avahi_dat$Density_km2 <- avahi_dat$Density*1000000

write.csv(avahi_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/avahi_dat.csv")

#Clean transect names
covs_avahi$Site_Transect[covs_avahi$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_avahi$Site_Transect[covs_avahi$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_avahi$Site_Transect[covs_avahi$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

#Join density data to covariate data 
covs_avahi_density <- left_join(avahi_dat, covs_avahi, by="Site_Transect")
covs_avahi_density$Species <- "Avahi_laniger"

#Linear model of density as explained by combinations of the covariates
lmer_avahi <- lmer(data=covs_avahi_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_avahi1 <- lmer(data=covs_avahi_density , Density~ elevation +(1|Site))
lmer_avahi2 <- lmer(data=covs_avahi_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_avahi3 <- lmer(data=covs_avahi_density , Density~  height + (1|Site))
lmer_avahi4 <- lmer(data=covs_avahi_density , Density~ rug +(1|Site))
lmer_avahi5 <- lmer(data=covs_avahi_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_avahi6 <- lmer(data=covs_avahi_density , Density~ elevation + height +(1|Site))
lmer_avahi7 <- lmer(data=covs_avahi_density , Density~ elevation +  rug +(1|Site))
lmer_avahi8 <- lmer(data=covs_avahi_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_avahi9 <- lmer(data=covs_avahi_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_avahi10 <- lmer(data=covs_avahi_density , Density~ height + rug +(1|Site))
lmer_avahi11 <- lmer(data=covs_avahi_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_avahi12 <- lmer(data=covs_avahi_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_avahi13 <- lmer(data=covs_avahi_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_avahi14 <- lmer(data=covs_avahi_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_avahi_rsq <- rsq.lmm(lmer_avahi)[[1]]
lmer_avahi_rsq1 <- rsq.lmm(lmer_avahi1)[[1]]
lmer_avahi_rsq2 <- rsq.lmm(lmer_avahi2)[[1]]
lmer_avahi_rsq3 <- rsq.lmm(lmer_avahi3)[[1]]
lmer_avahi_rsq4 <- rsq.lmm(lmer_avahi4)[[1]]
lmer_avahi_rsq5 <- rsq.lmm(lmer_avahi5)[[1]]
lmer_avahi_rsq6 <- rsq.lmm(lmer_avahi6)[[1]]
lmer_avahi_rsq7 <- rsq.lmm(lmer_avahi7)[[1]]
lmer_avahi_rsq8 <- rsq.lmm(lmer_avahi8)[[1]]
lmer_avahi_rsq9 <- rsq.lmm(lmer_avahi9)[[1]]
lmer_avahi_rsq10 <- rsq.lmm(lmer_avahi10)[[1]]
lmer_avahi_rsq11<- rsq.lmm(lmer_avahi11)[[1]]
lmer_avahi_rsq12 <- rsq.lmm(lmer_avahi12)[[1]]
lmer_avahi_rsq13 <- rsq.lmm(lmer_avahi13)[[1]]
lmer_avahi_rsq14 <- rsq.lmm(lmer_avahi14)[[1]]
max(lmer_avahi_rsq, lmer_avahi_rsq1 , lmer_avahi_rsq2 , lmer_avahi_rsq3 , lmer_avahi_rsq4 , lmer_avahi_rsq5 , lmer_avahi_rsq6, lmer_avahi_rsq7 , lmer_avahi_rsq8 , lmer_avahi_rsq9 , lmer_avahi_rsq10, lmer_avahi_rsq11 , lmer_avahi_rsq12 , lmer_avahi_rsq13 , lmer_avahi_rsq14 )

lmer_avahi_best <- lmer_avahi #rugedness has a significantly negative relationship with avahi density, height is significantly positive, and forest loss proportion is significantly positive
summary(lmer_avahi)

```

```{r Rubriventer density calculations}
#Get data frame for just rubriventer
rubriventer_df <- survey_diurnal  %>% filter(Nom_sientifique=="Eulemur_rubriventer") 

#Calculate perpendicular distance from angle and distance
#rubriventer_df$distance[is.na(rubriventer_df$distance)] <- 0

#ggplot(rubriventer_df, aes(distance,as.numeric(rubriventer_df$N_total)))+geom_point()+geom_smooth(method=lm)

#Clean data frame
rubriventer_df3 <- rubriventer_df
rubriventer_df3$Area <- 1000000 #For now, assume an area of 1... million 
rubriventer_df3$Sample.Label <- 1
rubriventer_df3$Region.Label <- rubriventer_df3$Site_Transect
rubriventer_df3$size <- as.numeric(rubriventer_df3$N_total)
rubriventer_df3$Effort <- rubriventer_df3$Effort.distance

#Detection function models
rubriventer_truncation_dist <- 30
ds_rubriventer <- ds(rubriventer_df3, truncation = rubriventer_truncation_dist)
gof_ds(ds_rubriventer) #significant, poor fit

model1_rubriventer <- ds(rubriventer_df3, key="hn", adjustment=NULL)
model2_rubriventer <- ds(rubriventer_df3, key="hn", adjustment="cos")
model3_rubriventer <- ds(rubriventer_df3, key="hn", adjustment="herm")
model4_rubriventer <- ds(rubriventer_df3, key="hr", adjustment=NULL)
model5_rubriventer <- ds(rubriventer_df3, key="hr", adjustment="cos")
model6_rubriventer <- ds(rubriventer_df3, key="hr", adjustment="herm")
model7_rubriventer <- ds(rubriventer_df3, key="unif", adjustment=NULL)
model8_rubriventer <- ds(rubriventer_df3, key="unif", adjustment="cos")
model9_rubriventer <- ds(rubriventer_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_rubriventer)#significant
gof_ds(model2_rubriventer)#significant
gof_ds(model3_rubriventer)#significant
gof_ds(model4_rubriventer)#significant
gof_ds(model5_rubriventer)#significant
gof_ds(model6_rubriventer)#significant
gof_ds(model7_rubriventer)#significant
gof_ds(model8_rubriventer)#significant
gof_ds(model9_rubriventer)#significant

#all significant :(

# compare AIC calues
AIC(model1_rubriventer, model2_rubriventer, model3_rubriventer,model4_rubriventer, model5_rubriventer, model6_rubriventer, model7_rubriventer,  model8_rubriventer ,model9_rubriventer)

#Plot the best-fitting model 
plot(model4_rubriventer, nc=5)
#4 looks ridiculous, so try 2 
plot(model2_rubriventer, nc=5)

#Get detection probability
rubriventer_detection_prob <- predict(ds_rubriventer)$fitted[[1]]

#Calculate strp width (2 * maximum distance, of the truncated distance data)
rubriventer_strip_width <- rubriventer_truncation_dist

#Upload covariate data 
rubriventer_df_old <- survey_all %>% filter(Nom_sientifique=="Eulemur_rubriventer" & Type=="Diurnal" & Date < ymd("20220601") & (Dis <1.56*rubriventer_truncation_dist | is.na(Dis)))
rubriventer_df_new <- survey_all %>% filter(Nom_sientifique=="Eulemur_rubriventer" & Type=="Diurnal"  & Date >= ymd("20220601") & (distance <= rubriventer_truncation_dist  | is.na(distance)) )
rubriventer_dat1 <- rbind(rubriventer_df_new, rubriventer_df_old)

rubriventer_dat <- rubriventer_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
rubriventer_dat <- left_join(site_transects, rubriventer_dat, by =c("Site_Transect"))

nobs_rubriventer <- rubriventer_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
rubriventer_dat <- left_join(rubriventer_dat, nobs_rubriventer, by =c("Site_Transect"))
rubriventer_dat[is.na(rubriventer_dat)] <- 0

rubriventer_dat <- left_join(rubriventer_dat,nrep_diurnal, by =c("Site_Transect"))
rubriventer_dat <- rubriventer_dat %>% select(-Site, -Transect, -Type)

rubriventer_dat <- left_join(rubriventer_dat,transect_lengths, by =c("Site_Transect"))
rubriventer_dat$Density <- (rubriventer_dat$Nind/rubriventer_detection_prob)/(rubriventer_truncation_dist*rubriventer_dat$Length*rubriventer_dat$NRep)
rubriventer_dat$Density_km2 <- rubriventer_dat$Density*1000000

write.csv(rubriventer_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/rubriventer_dat.csv")

#Clean transect names
covs_rubriventer$Site_Transect[covs_rubriventer$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_rubriventer$Site_Transect[covs_rubriventer$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_rubriventer$Site_Transect[covs_rubriventer$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

#Join density data to covariate data 
covs_rubriventer_density <- left_join(rubriventer_dat, covs_rubriventer, by="Site_Transect")
covs_rubriventer_density$Species <- "Eulemur_rubriventer"

#Linear model of density as explained by combinations of the covariates
lmer_rubriventer <- lmer(data=covs_rubri_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_rubriventer1 <- lmer(data=covs_rubri_density , Density~ elevation +(1|Site))
lmer_rubriventer2 <- lmer(data=covs_rubri_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_rubriventer3 <- lmer(data=covs_rubri_density , Density~  height + (1|Site))
lmer_rubriventer4 <- lmer(data=covs_rubri_density , Density~ rug +(1|Site))
lmer_rubriventer5 <- lmer(data=covs_rubri_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_rubriventer6 <- lmer(data=covs_rubri_density , Density~ elevation + height +(1|Site))
lmer_rubriventer7 <- lmer(data=covs_rubri_density , Density~ elevation +  rug +(1|Site))
lmer_rubriventer8 <- lmer(data=covs_rubri_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_rubriventer9 <- lmer(data=covs_rubri_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_rubriventer10 <- lmer(data=covs_rubri_density , Density~ height + rug +(1|Site))
lmer_rubriventer11 <- lmer(data=covs_rubri_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_rubriventer12 <- lmer(data=covs_rubri_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_rubriventer13 <- lmer(data=covs_rubri_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_rubriventer14 <- lmer(data=covs_rubri_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_rubriventer_rsq <- rsq.lmm(lmer_rubriventer)[[1]]
lmer_rubriventer_rsq1 <- rsq.lmm(lmer_rubriventer1)[[1]]
lmer_rubriventer_rsq2 <- rsq.lmm(lmer_rubriventer2)[[1]]
lmer_rubriventer_rsq3 <- rsq.lmm(lmer_rubriventer3)[[1]]
lmer_rubriventer_rsq4 <- rsq.lmm(lmer_rubriventer4)[[1]]
lmer_rubriventer_rsq5 <- rsq.lmm(lmer_rubriventer5)[[1]]
lmer_rubriventer_rsq6 <- rsq.lmm(lmer_rubriventer6)[[1]]
lmer_rubriventer_rsq7 <- rsq.lmm(lmer_rubriventer7)[[1]]
lmer_rubriventer_rsq8 <- rsq.lmm(lmer_rubriventer8)[[1]]
lmer_rubriventer_rsq9 <- rsq.lmm(lmer_rubriventer9)[[1]]
lmer_rubriventer_rsq10 <- rsq.lmm(lmer_rubriventer10)[[1]]
lmer_rubriventer_rsq11<- rsq.lmm(lmer_rubriventer11)[[1]]
lmer_rubriventer_rsq12 <- rsq.lmm(lmer_rubriventer12)[[1]]
lmer_rubriventer_rsq13 <- rsq.lmm(lmer_rubriventer13)[[1]]
lmer_rubriventer_rsq14 <- rsq.lmm(lmer_rubriventer14)[[1]]
max(lmer_rubriventer_rsq, lmer_rubriventer_rsq1 , lmer_rubriventer_rsq2 , lmer_rubriventer_rsq3 , lmer_rubriventer_rsq4 , lmer_rubriventer_rsq5 , lmer_rubriventer_rsq6, lmer_rubriventer_rsq7 , lmer_rubriventer_rsq8 , lmer_rubriventer_rsq9 , lmer_rubriventer_rsq10, lmer_rubriventer_rsq11 , lmer_rubriventer_rsq12 , lmer_rubriventer_rsq13 , lmer_rubriventer_rsq14 )

lmer_rubriventer_best <- lmer_rubriventer9 #Forest loss proportion is positively, significantly related to Eulemur rubriventer density 
summary(lmer_rubriventer_best)

```

```{r Albifrons density calculations}

#Get data frame for just albifrons
albifrons_df <- survey_diurnal  %>% filter(Nom_sientifique=="Eulemur_albifrons") 

#Calculate perpendicular distance from angle and distance
#albifrons_df$distance[is.na(albifrons_df$distance)] <- 0

#ggplot(albifrons_df, aes(distance,as.numeric(albifrons_df$N_total)))+geom_point()+geom_smooth(method=lm)

#Clean data frame
albifrons_df3 <- albifrons_df
albifrons_df3$Area <- 1000000 #For now, assume an area of 1... million 
albifrons_df3$Sample.Label <- 1
albifrons_df3$Region.Label <- albifrons_df3$Site_Transect
albifrons_df3$size <- as.numeric(albifrons_df3$N_total)
albifrons_df3$Effort <- albifrons_df3$Effort.distance

#Detection function models
albifrons_truncation_dist <- 25
ds_albifrons <- ds(albifrons_df3, truncation = albifrons_truncation_dist)
gof_ds(ds_albifrons)

model1_albifrons <- ds(albifrons_df3, key="hn", adjustment=NULL)
model2_albifrons <- ds(albifrons_df3, key="hn", adjustment="cos")
model3_albifrons <- ds(albifrons_df3, key="hn", adjustment="herm")
model4_albifrons <- ds(albifrons_df3, key="hr", adjustment=NULL)
model5_albifrons <- ds(albifrons_df3, key="hr", adjustment="cos")
model6_albifrons <- ds(albifrons_df3, key="hr", adjustment="herm")
model7_albifrons <- ds(albifrons_df3, key="unif", adjustment=NULL)
model8_albifrons <- ds(albifrons_df3, key="unif", adjustment="cos")
model9_albifrons <- ds(albifrons_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_albifrons)
gof_ds(model2_albifrons)
gof_ds(model3_albifrons)
gof_ds(model4_albifrons)
gof_ds(model5_albifrons)
gof_ds(model6_albifrons)
gof_ds(model7_albifrons)
gof_ds(model8_albifrons)
gof_ds(model9_albifrons)

#Of the models without poor fit, compare AIC calues
AIC(model1_albifrons, model2_albifrons, model3_albifrons,model4_albifrons,model5_albifrons,model6_albifrons,model7_albifrons, model8_albifrons, model9_albifrons)

#Plot the best-fitting model 
plot(model7_albifrons, nc=5)

#Get detection probability
albifrons_detection_prob <- predict(ds_albifrons)$fitted[[1]]

#Calculate strp width (2 * maximum distance, of the truncated distance data)
albifrons_strip_width <- albifrons_truncation_dist

#Upload covariate data 
albifrons_df_old <- survey_all %>% filter(Nom_sientifique=="Eulemur_albifrons" & Type=="Diurnal" & Date < ymd("20220601") & (Dis <1.56*albifrons_truncation_dist | is.na(Dis)))
albifrons_df_new <- survey_all %>% filter(Nom_sientifique=="Eulemur_albifrons" & Type=="Diurnal"  & Date >= ymd("20220601") & (distance <= albifrons_truncation_dist  | is.na(distance)) )
albi_dat1 <- rbind(albifrons_df_new, albifrons_df_old)

albi_dat <- albi_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
albi_dat <- left_join(site_transects, albi_dat, by =c("Site_Transect"))

nobs_albifrons <- albi_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
albi_dat <- left_join(albi_dat, nobs_albifrons, by =c("Site_Transect"))
albi_dat[is.na(albi_dat)] <- 0

albi_dat <- left_join(albi_dat,nrep_diurnal, by =c("Site_Transect"))
albi_dat <- albi_dat %>% select(-Site, -Transect, -Type)

albi_dat <- left_join(albi_dat,transect_lengths, by =c("Site_Transect"))
albi_dat$Density <- (albi_dat$Nind/albifrons_detection_prob)/(albifrons_truncation_dist*albi_dat$Length*albi_dat$NRep)
albi_dat$Density_km2 <- albi_dat$Density*1000000

write.csv(albi_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/albi_dat.csv")

covs_albi$Site_Transect[covs_albi$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_albi$Site_Transect[covs_albi$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_albi$Site_Transect[covs_albi$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

covs_albi_density <- left_join(albi_dat, covs_albi, by="Site_Transect")
covs_albi_density$Species <- "Eulemur_albifrons"

albi_ind <- survey_diurnal %>% filter(Nom_sientifique=="Eulemur_albifrons") %>% drop_na(N_ind)
albi_ind2 <- mean(as.numeric(albi_ind$N_ind))

#covs_albi_density$Density <- covs_albi_density$Density *albi_ind2

#Linear model of density as explained by combinations of the covariates
lmer_albifrons <- lmer(data=covs_albi_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_albifrons1 <- lmer(data=covs_albi_density , Density~ elevation +(1|Site))
lmer_albifrons2 <- lmer(data=covs_albi_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_albifrons3 <- lmer(data=covs_albi_density , Density~  height + (1|Site))
lmer_albifrons4 <- lmer(data=covs_albi_density , Density~ rug +(1|Site))
lmer_albifrons5 <- lmer(data=covs_albi_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_albifrons6 <- lmer(data=covs_albi_density , Density~ elevation + height +(1|Site))
lmer_albifrons7 <- lmer(data=covs_albi_density , Density~ elevation +  rug +(1|Site))
lmer_albifrons8 <- lmer(data=covs_albi_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_albifrons9 <- lmer(data=covs_albi_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_albifrons10 <- lmer(data=covs_albi_density , Density~ height + rug +(1|Site))
lmer_albifrons11 <- lmer(data=covs_albi_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_albifrons12 <- lmer(data=covs_albi_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_albifrons13 <- lmer(data=covs_albi_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_albifrons14 <- lmer(data=covs_albi_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_albifrons_rsq <- rsq.lmm(lmer_albifrons)[[1]]
lmer_albifrons_rsq1 <- rsq.lmm(lmer_albifrons1)[[1]]
lmer_albifrons_rsq2 <- rsq.lmm(lmer_albifrons2)[[1]]
lmer_albifrons_rsq3 <- rsq.lmm(lmer_albifrons3)[[1]]
lmer_albifrons_rsq4 <- rsq.lmm(lmer_albifrons4)[[1]]
lmer_albifrons_rsq5 <- rsq.lmm(lmer_albifrons5)[[1]]
lmer_albifrons_rsq6 <- rsq.lmm(lmer_albifrons6)[[1]]
lmer_albifrons_rsq7 <- rsq.lmm(lmer_albifrons7)[[1]]
lmer_albifrons_rsq8 <- rsq.lmm(lmer_albifrons8)[[1]]
lmer_albifrons_rsq9 <- rsq.lmm(lmer_albifrons9)[[1]]
lmer_albifrons_rsq10 <- rsq.lmm(lmer_albifrons10)[[1]]
lmer_albifrons_rsq11<- rsq.lmm(lmer_albifrons11)[[1]]
lmer_albifrons_rsq12 <- rsq.lmm(lmer_albifrons12)[[1]]
lmer_albifrons_rsq13 <- rsq.lmm(lmer_albifrons13)[[1]]
lmer_albifrons_rsq14 <- rsq.lmm(lmer_albifrons14)[[1]]
max(lmer_albifrons_rsq, lmer_albifrons_rsq1 , lmer_albifrons_rsq2 , lmer_albifrons_rsq3 , lmer_albifrons_rsq4 , lmer_albifrons_rsq5 , lmer_albifrons_rsq6, lmer_albifrons_rsq7 , lmer_albifrons_rsq8 , lmer_albifrons_rsq9 , lmer_albifrons_rsq10, lmer_albifrons_rsq11 , lmer_albifrons_rsq12 , lmer_albifrons_rsq13 , lmer_albifrons_rsq14 )

lmer_albifrons_best <-lmer_albifrons6
summary(lmer_albifrons6)

```

```{r Propithecus density calculations}

#Get data frame for just propithecus
propithecus_df <- survey_diurnal  %>% filter(Nom_sientifique=="Propithecus_candidus") 

#Calculate perpendicular distance from angle and distance
#propithecus_df$distance[is.na(propithecus_df$distance)] <- 0

#ggplot(propithecus_df, aes(distance,as.numeric(propithecus_df$N_total)))+geom_point()+geom_smooth(method=lm)

#Clean data frame
propithecus_df3 <- propithecus_df
propithecus_df3$Area <- 1000000 #For now, assume an area of 1... million 
propithecus_df3$Sample.Label <- 1
propithecus_df3$Region.Label <- propithecus_df3$Site_Transect
propithecus_df3$size <- as.numeric(propithecus_df3$N_total)
propithecus_df3$Effort <- propithecus_df3$Effort.distance

#Detection function models
  propithecus_truncation_dist <- 30
ds_propithecus <- ds(propithecus_df3, truncation = propithecus_truncation_dist)
gof_ds(ds_propithecus)

model1_propithecus <- ds(propithecus_df3, key="hn", adjustment=NULL)
model2_propithecus <- ds(propithecus_df3, key="hn", adjustment="cos")
model3_propithecus <- ds(propithecus_df3, key="hn", adjustment="herm")
model4_propithecus <- ds(propithecus_df3, key="hr", adjustment=NULL)
model5_propithecus <- ds(propithecus_df3, key="hr", adjustment="cos")
model6_propithecus <- ds(propithecus_df3, key="hr", adjustment="herm")#poor fit 
model7_propithecus <- ds(propithecus_df3, key="unif", adjustment=NULL)
model8_propithecus <- ds(propithecus_df3, key="unif", adjustment="cos")
model9_propithecus <- ds(propithecus_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_propithecus)
gof_ds(model2_propithecus)
gof_ds(model3_propithecus)
gof_ds(model4_propithecus)#significant
gof_ds(model5_propithecus)#significant
gof_ds(model6_propithecus)#significant
gof_ds(model7_propithecus)#significant
gof_ds(model8_propithecus)
gof_ds(model9_propithecus)#significant

#Of the models without poor fit, compare AIC calues
AIC(model1_propithecus,model2_propithecus,model3_propithecus, model8_propithecus)

#Plot the best-fitting model 
plot(model2_propithecus, nc=5)

#Get detection probability
propithecus_detection_prob <- predict(ds_propithecus)$fitted[[1]]

#Calculate strp width (2 * maximum distance, of the truncated distance data)
propithecus_strip_width <- propithecus_truncation_dist

#Upload covariate data 
propithecus_df_old <- survey_all %>% filter(Nom_sientifique=="Propithecus_candidus" & Type=="Diurnal" & Date < ymd("20220601") & (Dis <1.56*propithecus_truncation_dist | is.na(Dis)))
propithecus_df_new <- survey_all %>% filter(Nom_sientifique=="Propithecus_candidus" & Type=="Diurnal"  & Date >= ymd("20220601") & (distance <= propithecus_truncation_dist  | is.na(distance)) )
prop_dat1 <- rbind(propithecus_df_new, propithecus_df_old)

prop_dat <- prop_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
prop_dat <- left_join(site_transects, prop_dat, by =c("Site_Transect"))

nobs_propithecus <- prop_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
prop_dat <- left_join(prop_dat, nobs_propithecus, by =c("Site_Transect"))
prop_dat[is.na(prop_dat)] <- 0

prop_dat <- left_join(prop_dat,nrep_diurnal, by =c("Site_Transect"))
prop_dat <- prop_dat %>% select(-Site, -Transect, -Type)

prop_dat <- left_join(prop_dat,transect_lengths, by =c("Site_Transect"))
prop_dat$Density <- (prop_dat$Nind/propithecus_detection_prob)/(propithecus_truncation_dist*prop_dat$Length*prop_dat$NRep)
prop_dat$Density_km2 <- prop_dat$Density*1000000

write.csv(prop_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/prop_dat.csv")

covs_prop$Site_Transect[covs_prop$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_prop$Site_Transect[covs_prop$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_prop$Site_Transect[covs_prop$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

covs_prop_density <- left_join(prop_dat, covs_prop, by="Site_Transect")
covs_prop_density$Species <- "Propithecus_candidus"

prop_ind <- survey_diurnal %>% filter(Nom_sientifique=="Propithecus_candidus") %>% drop_na(N_ind)
prop_ind2 <- mean(as.numeric(prop_ind$N_ind))

covs_prop_density$Density <- covs_prop_density$Density *prop_ind2

#Linear model of density as explained by combinations of the covariates
lmer_propithecus <- lmer(data=covs_prop_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_propithecus1 <- lmer(data=covs_prop_density , Density~ elevation +(1|Site))
lmer_propithecus2 <- lmer(data=covs_prop_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_propithecus3 <- lmer(data=covs_prop_density , Density~  height + (1|Site))
lmer_propithecus4 <- lmer(data=covs_prop_density , Density~ rug +(1|Site))
lmer_propithecus5 <- lmer(data=covs_prop_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_propithecus6 <- lmer(data=covs_prop_density , Density~ elevation + height +(1|Site))
lmer_propithecus7 <- lmer(data=covs_prop_density , Density~ elevation +  rug +(1|Site))
lmer_propithecus8 <- lmer(data=covs_prop_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_propithecus9 <- lmer(data=covs_prop_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_propithecus10 <- lmer(data=covs_prop_density , Density~ height + rug +(1|Site))
lmer_propithecus11 <- lmer(data=covs_prop_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_propithecus12 <- lmer(data=covs_prop_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_propithecus13 <- lmer(data=covs_prop_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_propithecus14 <- lmer(data=covs_prop_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_propithecus_rsq <- rsq.lmm(lmer_propithecus)[[1]]
lmer_propithecus_rsq1 <- rsq.lmm(lmer_propithecus1)[[1]]
lmer_propithecus_rsq2 <- rsq.lmm(lmer_propithecus2)[[1]]
lmer_propithecus_rsq3 <- rsq.lmm(lmer_propithecus3)[[1]]
lmer_propithecus_rsq4 <- rsq.lmm(lmer_propithecus4)[[1]]
lmer_propithecus_rsq5 <- rsq.lmm(lmer_propithecus5)[[1]]
lmer_propithecus_rsq6 <- rsq.lmm(lmer_propithecus6)[[1]]
lmer_propithecus_rsq7 <- rsq.lmm(lmer_propithecus7)[[1]]
lmer_propithecus_rsq8 <- rsq.lmm(lmer_propithecus8)[[1]]
lmer_propithecus_rsq9 <- rsq.lmm(lmer_propithecus9)[[1]]
lmer_propithecus_rsq10 <- rsq.lmm(lmer_propithecus10)[[1]]
lmer_propithecus_rsq11<- rsq.lmm(lmer_propithecus11)[[1]]
lmer_propithecus_rsq12 <- rsq.lmm(lmer_propithecus12)[[1]]
lmer_propithecus_rsq13 <- rsq.lmm(lmer_propithecus13)[[1]]
lmer_propithecus_rsq14 <- rsq.lmm(lmer_propithecus14)[[1]]
max(lmer_propithecus_rsq, lmer_propithecus_rsq1 , lmer_propithecus_rsq2 , lmer_propithecus_rsq3 , lmer_propithecus_rsq4 , lmer_propithecus_rsq5 , lmer_propithecus_rsq6, lmer_propithecus_rsq7 , lmer_propithecus_rsq8 , lmer_propithecus_rsq9 , lmer_propithecus_rsq10, lmer_propithecus_rsq11 , lmer_propithecus_rsq12 , lmer_propithecus_rsq13 , lmer_propithecus_rsq14 )

lmer_propithecus_best <- lmer_propithecus10
summary(lmer_propithecus_best)

```

```{r Hapalemur density calculations}
#Get data frame for just hapalemur
hapalemur_df <- survey_diurnal  %>% filter(Nom_sientifique=="Hapalemur_occidentalis") 

#Calculate perpendicular distance from angle and distance
#hapalemur_df$distance[is.na(hapalemur_df$distance)] <- 0

#Clean data frame
hapalemur_df3 <- hapalemur_df
hapalemur_df3$Area <- 1000000 #For now, assume an area of 1... million 
hapalemur_df3$Sample.Label <- 1
hapalemur_df3$Region.Label <- hapalemur_df3$Site_Transect
hapalemur_df3$size <- as.numeric(hapalemur_df3$N_total)
hapalemur_df3$Effort <- hapalemur_df3$Effort.distance


#Detection function models
hapalemur_truncation_dist <- 18
ds_hapalemur <- ds(hapalemur_df3, truncation = hapalemur_truncation_dist)
gof_ds(ds_hapalemur)

model1_hapalemur <- ds(hapalemur_df3, key="hn", adjustment=NULL)
model2_hapalemur <- ds(hapalemur_df3, key="hn", adjustment="cos")
model3_hapalemur <- ds(hapalemur_df3, key="hn", adjustment="herm")
model4_hapalemur <- ds(hapalemur_df3, key="hr", adjustment=NULL)
model5_hapalemur <- ds(hapalemur_df3, key="hr", adjustment="cos")
model6_hapalemur <- ds(hapalemur_df3, key="hr", adjustment="herm")
model7_hapalemur <- ds(hapalemur_df3, key="unif", adjustment=NULL)
model8_hapalemur <- ds(hapalemur_df3, key="unif", adjustment="cos")
model9_hapalemur <- ds(hapalemur_df3, key="unif", adjustment="herm")

#Test goodness of fit for detection functions
gof_ds(model1_hapalemur)
gof_ds(model2_hapalemur)
gof_ds(model3_hapalemur)
gof_ds(model4_hapalemur)
gof_ds(model5_hapalemur)
gof_ds(model6_hapalemur)
gof_ds(model7_hapalemur)
gof_ds(model8_hapalemur)
gof_ds(model9_hapalemur)

#Of the models without poor fit, compare AIC calues
AIC(model1_hapalemur, model2_hapalemur, model3_hapalemur, model4_hapalemur,model5_hapalemur,model6_hapalemur,model7_hapalemur,model8_hapalemur, model9_hapalemur)

#Plot the best-fitting model 
plot(model7_hapalemur, nc=5)

#Get detection probability
hapalemur_detection_prob <- predict(ds_hapalemur)$fitted[[1]] #odd that this is 1? Assumed perfect detection but they are so cryptic! 

#Calculate strp width (2 * maximum distance, of the truncated distance data)
hapalemur_strip_width <- hapalemur_truncation_dist

#Upload covariate data 
hapalemur_df_old <- survey_all %>% filter(Nom_sientifique=="Hapalemur_occidentalis" & Type=="Diurnal" & Date < ymd("20220601") & (Dis <1.56*hapalemur_truncation_dist | is.na(Dis)))
hapalemur_df_new <- survey_all %>% filter(Nom_sientifique=="Hapalemur_occidentalis" & Type=="Diurnal"  & Date >= ymd("20220601") & (distance <= hapalemur_truncation_dist  | is.na(distance)) )
hapa_dat1 <- rbind(hapalemur_df_new, hapalemur_df_old)

hapa_dat <- hapa_dat1  %>% group_by(Site_Transect)%>% summarise(Nind = sum(as.numeric(N_total)))
hapa_dat <- left_join(site_transects, hapa_dat, by =c("Site_Transect"))

nobs_hapalemur <- hapa_dat1  %>% select(Site_Transect) %>%group_by(Site_Transect) %>% mutate(NObs =n())%>% distinct()
hapa_dat <- left_join(hapa_dat, nobs_hapalemur, by =c("Site_Transect"))
hapa_dat[is.na(hapa_dat)] <- 0

hapa_dat <- left_join(hapa_dat,nrep_diurnal, by =c("Site_Transect"))
hapa_dat <- hapa_dat %>% select(-Site, -Transect, -Type)

hapa_dat <- left_join(hapa_dat,transect_lengths, by =c("Site_Transect"))
hapa_dat$Density <- (hapa_dat$Nind/hapalemur_detection_prob)/(hapalemur_truncation_dist*hapa_dat$Length*hapa_dat$NRep)
hapa_dat$Density_km2 <- hapa_dat$Density*1000000

write.csv(hapa_dat,"~/Documents/GitHub/COMATSA_GitHub/COMATSA_GitHub/Data/prop_dat.csv")

covs_hapa$Site_Transect[covs_hapa$Site_Transect=="Ambodivohitra_Kobahina_1"]<- "AmbodivohitraKobahina_1"
covs_hapa$Site_Transect[covs_hapa$Site_Transect=="Ambodivohitra_Kobahina_2"]<- "AmbodivohitraKobahina_2"
covs_hapa$Site_Transect[covs_hapa$Site_Transect=="Ambodivohitra_Kobahina_3"]<- "AmbodivohitraKobahina_3"

covs_hapa_density <- left_join(hapa_dat, covs_hapa, by="Site_Transect")
covs_hapa_density$Species <- "Hapalemur_occidentalis"

hapa_ind <- survey_diurnal %>% filter(Nom_sientifique=="Hapalemur_occidentalis") %>% drop_na(N_ind)
hapa_ind2 <- mean(as.numeric(hapa_ind$N_ind))

covs_hapa_density$Density <- covs_hapa_density$Density *hapa_ind2

#Linear model of density as explained by combinations of the covariates
lmer_hapalemur <- lmer(data=covs_hapa_density , Density~ elevation + log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_hapalemur1 <- lmer(data=covs_hapa_density , Density~ elevation +(1|Site))
lmer_hapalemur2 <- lmer(data=covs_hapa_density , Density~  log(forest_loss_prop+1) +(1|Site))
lmer_hapalemur3 <- lmer(data=covs_hapa_density , Density~  height + (1|Site))
lmer_hapalemur4 <- lmer(data=covs_hapa_density , Density~ rug +(1|Site))
lmer_hapalemur5 <- lmer(data=covs_hapa_density , Density~ elevation + log(forest_loss_prop+1) + (1|Site))
lmer_hapalemur6 <- lmer(data=covs_hapa_density , Density~ elevation + height +(1|Site))
lmer_hapalemur7 <- lmer(data=covs_hapa_density , Density~ elevation +  rug +(1|Site))
lmer_hapalemur8 <- lmer(data=covs_hapa_density , Density~  log(forest_loss_prop+1) + height +(1|Site))
lmer_hapalemur9 <- lmer(data=covs_hapa_density , Density~ log(forest_loss_prop+1) +  rug +(1|Site))
lmer_hapalemur10 <- lmer(data=covs_hapa_density , Density~ height + rug +(1|Site))
lmer_hapalemur11 <- lmer(data=covs_hapa_density , Density~ elevation + log(forest_loss_prop+1) + height +(1|Site))
lmer_hapalemur12 <- lmer(data=covs_hapa_density , Density~ elevation + log(forest_loss_prop+1) +  rug +(1|Site))
lmer_hapalemur13 <- lmer(data=covs_hapa_density , Density~  log(forest_loss_prop+1) + height + rug +(1|Site))
lmer_hapalemur14 <- lmer(data=covs_hapa_density , Density~  +(1|Site))

#Check for the model with the highest r-squared
lmer_hapalemur_rsq <- rsq.lmm(lmer_hapalemur)[[1]]
lmer_hapalemur_rsq1 <- rsq.lmm(lmer_hapalemur1)[[1]]
lmer_hapalemur_rsq2 <- rsq.lmm(lmer_hapalemur2)[[1]]
lmer_hapalemur_rsq3 <- rsq.lmm(lmer_hapalemur3)[[1]]
lmer_hapalemur_rsq4 <- rsq.lmm(lmer_hapalemur4)[[1]]
lmer_hapalemur_rsq5 <- rsq.lmm(lmer_hapalemur5)[[1]]
lmer_hapalemur_rsq6 <- rsq.lmm(lmer_hapalemur6)[[1]]
lmer_hapalemur_rsq7 <- rsq.lmm(lmer_hapalemur7)[[1]]
lmer_hapalemur_rsq8 <- rsq.lmm(lmer_hapalemur8)[[1]]
lmer_hapalemur_rsq9 <- rsq.lmm(lmer_hapalemur9)[[1]]
lmer_hapalemur_rsq10 <- rsq.lmm(lmer_hapalemur10)[[1]]
lmer_hapalemur_rsq11<- rsq.lmm(lmer_hapalemur11)[[1]]
lmer_hapalemur_rsq12 <- rsq.lmm(lmer_hapalemur12)[[1]]
lmer_hapalemur_rsq13 <- rsq.lmm(lmer_hapalemur13)[[1]]
lmer_hapalemur_rsq14 <- rsq.lmm(lmer_hapalemur14)[[1]]
max(lmer_hapalemur_rsq, lmer_hapalemur_rsq1 , lmer_hapalemur_rsq2 , lmer_hapalemur_rsq3 , lmer_hapalemur_rsq4 , lmer_hapalemur_rsq5 , lmer_hapalemur_rsq6, lmer_hapalemur_rsq7 , lmer_hapalemur_rsq8 , lmer_hapalemur_rsq9 , lmer_hapalemur_rsq10, lmer_hapalemur_rsq11 , lmer_hapalemur_rsq12 , lmer_hapalemur_rsq13 , lmer_hapalemur_rsq14 )

lmer_hapalemur_best <- lmer_hapalemur
summary(lmer_hapalemur_best)

```

```{r Co-occurence}
library(cooccur)
data(finches)
finches

cooccur_dat <- survey_all %>% select(Site, Transect, Nom_sientifique) %>% unite("Site_Transect", Site:Transect)%>% distinct()
cooccur_dat$occur <- 1
cooccur_dat$occur[is.na(cooccur_dat$Nom_sientifique)] <- 0
cooccur_dat <- cooccur_dat[-44,]
cooccur_dat <- cooccur_dat[!is.na(cooccur_dat$Nom_sientifique),]
cooccur_dat <- cooccur_dat %>% pivot_wider(names_from = "Site_Transect", values_from= "occur")
cooccur_dat[is.na(cooccur_dat)] <- 0
cooccur_dat <- as.data.frame(cooccur_dat)
rownames(cooccur_dat) <- cooccur_dat$Nom_sientifique
cooccur_dat <- cooccur_dat[,-1]

cooccur_lemurs <- cooccur(mat=cooccur_dat, thresh = FALSE, type="spp_site", spp_names = TRUE)
summary(cooccur_lemurs)
plot(cooccur_lemurs, plotrand = TRUE)

class(cooccur.finches)

```

```{r Figures}

#heatmap figure
density_all <- rbind(covs_albi_density, covs_allo_density, covs_cheiro_density, covs_micro_density, covs_avahi_density, covs_lepi_density, covs_rubri_density, covs_hapa_density, covs_prop_density)

library(viridis)
ggplot(data=density_all, aes(x= Site_Transect, y=Species, fill=log(Density+1)))+
  geom_tile()+
  scale_fill_viridis(name="log(Density (individuals/km2))")+
  theme_classic()+ 
  xlab("Site and Transect")+
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

# Scatter plots for significant density relationships

forest_avahi_plot <- ggplot(data=covs_avahi_density, aes(x=log(forest_loss_prop+1), y=Density))+
  geom_point()+
  geom_smooth(method="lm", color="darkorange")+
  theme_classic()+
  ylab("Avahi density")+
  xlab("log (Forest Loss Proportion)")

height_avahi_plot <- ggplot(data=covs_avahi_density, aes(x=height, y=Density))+
  geom_point()+
  geom_smooth(method="lm")+
  theme_classic()+
  ylab("Avahi density")+
  xlab("Forest Height (m)")

elevation_avahi_plot <- ggplot(data=covs_avahi_density, aes(x=elevation, y=Density))+
  geom_point()+
  geom_smooth(method="lm")+
  theme_classic()+
  ylab("Avahi density")+
  xlab("Elevation (m)")

rug_avahi_plot <- ggplot(data=covs_avahi_density, aes(x=rug, y=Density))+
  geom_point()+
  geom_smooth(method="lm")+
  theme_classic()+
  ylab("Avahi density")+
  xlab("Ruggedness")

forest_cheiro_plot <- ggplot(data=covs_cheiro_density, aes(x=log(forest_loss_prop+1), y=Density))+
  geom_point()+
  geom_smooth(method="lm")+
  theme_classic()+
  ylab("Cheirogaleus density")+
  xlab("log (Forest Loss Proportion)")

forest_rubri_plot <- ggplot(data=covs_rubri_density, aes(x=log(forest_loss_prop+1), y=Density))+
  geom_point()+
  geom_smooth(method="lm")+
  theme_classic()+
  ylab("E. rubriventer density")+
  xlab("log (Forest Loss Proportion)")

forest_lepi_plot <- ggplot(data=covs_lepi_density, aes(x=rug, y=Density))+
  geom_point()+
  geom_smooth(method="lm")+
  theme_classic()+
  ylab("Lepilemur density")+
  xlab("Ruggedness")

library(ggpubr)

density_habitat_plot <- ggarrange(forest_cheiro_plot, forest_rubri_plot, forest_avahi_plot, rug_avahi_plot, height_avahi_plot, forest_lepi_plot,  labels= c("a", "b", "c", "d", "e", "f"), nrow=3, ncol=2)


```
